+++ apps/backend/src/main/kotlin/ca/uvictrades/trades/controller/trade/requests/PlaceStockOrderRequest.kt
@@ -1,13 +1,14 @@
 package ca.uvictrades.trades.controller.trade.requests
 
 import java.math.BigDecimal
+import java.math.BigInteger
 
 data class PlaceStockOrderRequest(
-    val stock_id: String,
-    val is_buy: Boolean,
-    val order_type: StockOrderType,
-    val quantity: Int,
-    val price: BigDecimal?,
+	val stock_id: String,
+	val is_buy: Boolean,
+	val order_type: StockOrderType,
+	val quantity: BigInteger,
+	val price: BigDecimal?,
 ) {
 	enum class StockOrderType {
 		LIMIT,
